chore(urls): Apply code review discussion (#3382)
