+++ hawtio-system/src/main/java/io/hawt/web/auth/AuthenticationFilter.java
@@ -3,12 +3,10 @@ package io.hawt.web.auth;
 import java.io.IOException;
 import java.security.PrivilegedActionException;
 import java.security.PrivilegedExceptionAction;
-import java.util.Collections;
 
 import javax.security.auth.Subject;
-import javax.security.auth.login.AppConfigurationEntry;
 
-import io.hawt.util.Strings;
+import io.hawt.util.WebHelper;
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.FilterConfig;
@@ -43,7 +41,7 @@ public class AuthenticationFilter implements Filter {
     public void init(FilterConfig filterConfig) throws ServletException {
         authConfiguration = AuthenticationConfiguration.getConfiguration(filterConfig.getServletContext());
         timeout = AuthSessionHelpers.getSessionTimeout(filterConfig.getServletContext());
-        this.pathIndex = Strings.hawtioPathIndex(filterConfig.getServletContext());
+        this.pathIndex = WebHelper.hawtioPathIndex(filterConfig.getServletContext());
     }
 
     @Override
