+++ hawtio-system/src/main/java/io/hawt/web/auth/RelativeRequestUri.java
@@ -2,9 +2,8 @@ package io.hawt.web.auth;
 
 import java.util.regex.Pattern;
 
+import io.hawt.util.WebHelper;
 import jakarta.servlet.http.HttpServletRequest;
-
-import io.hawt.util.Strings;
 
 /**
  * URI path relative to a given index. An index represents the number of path
@@ -62,7 +61,7 @@ public class RelativeRequestUri {
             throw new IllegalArgumentException("pathIndex is negative");
         }
 
-        final String requestUri = Strings.webContextPath(request.getRequestURI());
+        final String requestUri = WebHelper.webContextPath(request.getRequestURI());
         int start = request.getContextPath().length();
         if (start < requestUri.length() && requestUri.charAt(start) == '/') {
             start++;
