+++ hawtio-system/src/test/java/io/hawt/util/StringsTest.java
@@ -3,24 +3,14 @@ package io.hawt.util;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Properties;
-import java.util.stream.Stream;
 
-import jakarta.servlet.ServletContext;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
-import static io.hawt.web.auth.SessionExpiryFilter.SERVLET_PATH;
-import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.fail;
-import static org.junit.jupiter.params.provider.Arguments.arguments;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 
 public class StringsTest {
@@ -119,138 +109,4 @@ public class StringsTest {
         }
     }
 
-    public static class CleanPathTest {
-
-        public static Stream<Arguments> params() {
-            return Stream.of(
-                arguments("", ""),
-                arguments("  ", "  "),
-                arguments("/", "/"),
-                arguments("a", "a"),
-                arguments("/a", "/a"),
-                arguments("a/", "a"),
-                arguments("/a/", "/a"),
-                arguments("//a/", "/a"),
-                arguments("/a//", "/a"),
-                arguments("//a//", "/a"),
-                arguments("/a/b/", "/a/b"),
-                arguments("/a///b/", "/a/b")
-            );
-        }
-
-        @ParameterizedTest
-        @MethodSource("params")
-        public void test(String input, String expected) {
-            assertThat(Strings.cleanPath(input), equalTo(expected));
-        }
-    }
-
-
-    public static class WebContextPathFromSingleComponentTest {
-
-        public static Stream<Arguments> params() {
-            return Stream.of(
-                arguments(null, ""),
-                arguments("", ""),
-                arguments(" ", "/ "),
-                arguments("/", ""),
-                arguments("a", "/a"),
-                arguments("/a", "/a"),
-                arguments("a/", "/a"),
-                arguments("/a/", "/a"),
-                arguments("//a/", "/a"),
-                arguments("/a//", "/a"),
-                arguments("//a//", "/a"),
-                arguments("/a/b/", "/a/b"),
-                arguments("/a///b/", "/a/b")
-            );
-        }
-
-        @ParameterizedTest
-        @MethodSource("params")
-        public void test(String input, String expected) {
-            assertThat(Strings.webContextPath((input)), equalTo(expected));
-        }
-    }
-
-
-    public static class WebContextPathFromMultipleComponentsTest {
-
-        public static class Parameters {
-            private final String input;
-            private final String expected;
-            private final String more;
-
-            private Parameters(String input, String more, String expected) {
-                this.input = input;
-                this.expected = expected;
-                this.more = more;
-            }
-
-            public Parameters(String input, String expected) {
-                this.input = input;
-                this.expected = expected;
-                this.more = null;
-            }
-        }
-
-        public static Stream<Parameters> params() {
-            return Stream.of(
-                new Parameters(null, ""),
-                new Parameters("", ""),
-                new Parameters(" ", "/ "),
-                new Parameters("/", ""),
-                new Parameters("a", "/a"),
-                new Parameters("/a", "/a"),
-                new Parameters("a/", "/a"),
-                new Parameters("/a/", "/a"),
-                new Parameters("//a/", "/a"),
-                new Parameters("/a//", "/a"),
-                new Parameters("//a//", "/a"),
-                new Parameters(null, null, ""),
-                new Parameters(null, "a", "/a"),
-                new Parameters("a", null, "/a"),
-                new Parameters("a", "b", "/a/b"),
-                new Parameters("/a", "b", "/a/b"),
-                new Parameters("a", "/b", "/a/b"),
-                new Parameters("/a", "/b", "/a/b"),
-                new Parameters("/a/", "b", "/a/b"),
-                new Parameters("/a/", "/b", "/a/b"),
-                new Parameters("/a/", "/b/", "/a/b"),
-                new Parameters("/a//", "/b//", "/a/b")
-            );
-        }
-
-
-        @ParameterizedTest
-        @MethodSource("params")
-        public void test(Parameters args) {
-            assertThat(Strings.webContextPath(args.input, args.more), equalTo(args.expected));
-        }
-    }
-
-    public static class HawtioPathIndexTest {
-
-        public static Stream<Arguments> params() {
-            return Stream.of(
-                    // [ SERVLET_PATH attribute, full request URI, expected Hawtio path position ]
-                    arguments(null, "/jolokia", 0),
-                    arguments(null, "/jolokia/version", 0),
-                    arguments("", "/jolokia", 0),
-                    arguments("", "/jolokia/version", 0),
-                    arguments("/x", "/x/jolokia", 1),
-                    arguments("/mgmt/actuator/hawtio", "/mgmt/actuator/hawtio/jolokia", 3)
-            );
-        }
-
-        @ParameterizedTest
-        @MethodSource("params")
-        public void test(String servletPathAttribute, String uri, int pathIndex) {
-            ServletContext ctx = mock(ServletContext.class);
-            if (servletPathAttribute != null) {
-                when(ctx.getAttribute(SERVLET_PATH)).thenReturn(servletPathAttribute);
-            }
-            assertThat(Strings.hawtioPathIndex(ctx), equalTo(pathIndex));
-        }
-    }
 }
